pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "main", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.1"}
alembic = "1.13.2"
psycopg2 = "2.9.9"
pytest = "8.2.2"
pillow = "10.4.0"
opencv-python = "4.10.0.84"
sqlalchemy = "2.0.31"
authlib = "1.3.1"
transformers = {version = "4.42.4", extras = ["torch"]}
accelerate = "0.32.1"
boto3 = "1.34.147"
tensorflow = "2.16.2"
celery = "^5.4.0"
flower = "^2.0.1"
redis = "^5.0.8"
sse-starlette = "^2.1.3"
wget = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
